moves journey utils out to it's own util file
