+++ src/routes/TaskDetails/components/airpax/AirpaxTaskSummary.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { MovementUtil } from '../../../../utils';
+import { JourneyUtil, MovementUtil } from '../../../../utils';
 
 import { ICON } from '../../../../utils/constants';
 
 const AirpaxTaskSummary = ({ version, refDataAirlineCodes }) => {
-  const journey = MovementUtil.movementJourney(version);
+  const journey = JourneyUtil.get(version);
   const flight = MovementUtil.movementFlight(version);
-  const arrivalTime = MovementUtil.arrivalTime(journey);
+  const arrivalTime = JourneyUtil.arrivalTime(journey);
   return (
     <section className="task-list--voyage-section overflow-hidden">
       <div className="govuk-grid-row grid-background--greyed">
@@ -25,18 +25,18 @@ const AirpaxTaskSummary = ({ version, refDataAirlineCodes }) => {
           <i className="c-icon-aircraft" />
           <p className="content-line-one govuk-!-padding-right-2">
             {`${MovementUtil.airlineName(MovementUtil.airlineOperator(flight), refDataAirlineCodes)} flight, 
-              ${MovementUtil.voyageText(arrivalTime, true, MovementUtil.iataToCity(MovementUtil.arrivalLoc(journey)))}
+              ${MovementUtil.voyageText(arrivalTime, true, MovementUtil.iataToCity(JourneyUtil.arrivalLoc(journey)))}
               `}
           </p>
           <p className="govuk-body-s content-line-two govuk-!-padding-right-2">
             <span className="govuk-!-font-weight-bold">{MovementUtil.flightNumber(flight)}</span>
             <span className="dot" />
-            {`${MovementUtil.formatDepartureTime(journey)}`}
+            {`${JourneyUtil.formatDepartureTime(journey)}`}
             <span className="dot" />
-            <span className="govuk-!-font-weight-bold">{MovementUtil.departureLoc(journey)}</span> &#8594;
-            <span className="govuk-!-font-weight-bold"> {MovementUtil.arrivalLoc(journey)}</span>
+            <span className="govuk-!-font-weight-bold">{JourneyUtil.departureLoc(journey)}</span> &#8594;
+            <span className="govuk-!-font-weight-bold"> {JourneyUtil.arrivalLoc(journey)}</span>
             <span className="dot" />
-            {MovementUtil.formatArrivalTime(journey)}
+            {JourneyUtil.formatArrivalTime(journey)}
           </p>
         </div>
       </div>
