+++ src/routes/TaskDetails/components/airpax/Itinerary.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
 import { DATE_FORMATS, STRINGS } from '../../../../utils/constants';
-import { MovementUtil } from '../../../../utils';
+import { JourneyUtil, MovementUtil } from '../../../../utils';
 
 const Itinerary = ({ version }) => {
-  const itinerary = MovementUtil.movementItinerary(MovementUtil.movementJourney(version));
+  const itinerary = JourneyUtil.movementItinerary(JourneyUtil.get(version));
   return (
     <div className="task-details-container">
       <h3 className="govuk-heading-m govuk-!-margin-top-0">Itinerary</h3>
@@ -18,14 +18,14 @@ const Itinerary = ({ version }) => {
             {index !== 0 && index <= itinerary.length - 1 && MovementUtil.itinRelativeTime(index, it, itinerary)}
             <div className="font__bold">
               {MovementUtil.itinFlightNumber(it)} <span className="dot__light" />&nbsp;
-              {MovementUtil.departureLoc(it)} <span className="right-arrow">&#8594;</span>&nbsp;
-              {MovementUtil.arrivalLoc(it)} <span className="dot__light" />&nbsp;
-              {MovementUtil.formatDepartureTime(it, DATE_FORMATS.LONG_DAY_DATE)}
+              {JourneyUtil.departureLoc(it)} <span className="right-arrow">&#8594;</span>&nbsp;
+              {JourneyUtil.arrivalLoc(it)} <span className="dot__light" />&nbsp;
+              {JourneyUtil.formatDepartureTime(it, DATE_FORMATS.LONG_DAY_DATE)}
             </div>
             <div className="font__light">
               {MovementUtil.itinDepartureCountryCode(it)} <span className="right-arrow">&#8594;</span>&nbsp;
               {MovementUtil.itinArrivalCountryCode(it)} <span className="dot__light" />&nbsp;
-              {STRINGS.ARRIVAL_TEXT} {MovementUtil.formatArrivalTime(it, DATE_FORMATS.LONG)}
+              {STRINGS.ARRIVAL_TEXT} {JourneyUtil.formatArrivalTime(it, DATE_FORMATS.LONG)}
             </div>
           </div>
         );
