+++ src/utils/Movement/movementUtil.jsx
@@ -4,7 +4,6 @@ import lookup from 'country-code-lookup';
 import { Tag } from '@ukhomeoffice/cop-react-components';
 
 import {
-  DATE_FORMATS,
   DEPARTURE_STATUS,
   STRINGS,
   ICON,
@@ -12,17 +11,16 @@ import {
   MOVEMENT_MODES,
   MOVEMENT_ROLE,
   TASK_STATUS,
-  UNKNOWN_TIME_DATA, TASK_OUTCOME,
+  TASK_OUTCOME,
 } from '../constants';
 
 import { calculateTimeDifference,
-  getFormattedDate,
   isInPast,
   toDateTimeList,
   toRelativeTime } from '../Datetime/datetimeUtil';
 import { getTotalNumberOfPersons } from '../Person/personUtil';
-import { isNotNumber } from '../Number/numberUtil';
 import CommonUtil from '../Common/commonUtil';
+import JourneyUtil from '../Journey/journeyUtil';
 
 const getOccupantCounts = (targetTask) => {
   return targetTask?.movement?.occupants || undefined;
@@ -228,10 +226,6 @@ const getMovementModeIcon = (movementMode, vehicle, passengers) => {
   return ICON.NONE;
 };
 
-const getMovementDirection = (journey) => {
-  return journey?.direction || undefined;
-};
-
 const toVoyageText = (dateTime, isTaskDetails = false, prefix = '') => {
   const time = toRelativeTime(dateTime);
   const isPastDate = isInPast(dateTime);
@@ -274,55 +268,6 @@ const getItineraryFlightNumber = (itinerary) => {
   return itinerary.id;
 };
 
-const hasItinerary = (journey) => {
-  return !!journey?.itinerary;
-};
-
-const getMovementItinerary = (journey) => {
-  if (hasItinerary(journey)) {
-    return journey.itinerary;
-  }
-  return null;
-};
-
-const getFlightTimeObject = (milliseconds) => {
-  if (!milliseconds && milliseconds !== 0) {
-    return UNKNOWN_TIME_DATA;
-  }
-
-  if (isNotNumber(milliseconds)) {
-    return UNKNOWN_TIME_DATA;
-  }
-
-  let seconds = Math.floor(milliseconds / 1000);
-  let minutes = Math.floor(seconds / 60);
-  let hours = Math.floor(minutes / 60);
-
-  seconds %= 60;
-  minutes %= 60;
-
-  return { h: hours, m: minutes, s: seconds };
-};
-
-const getJourneyDuration = (journey) => {
-  if (!journey?.duration) {
-    return STRINGS.UNKNOWN_TEXT;
-  }
-  return journey.duration;
-};
-
-const toFormattedFlightTime = (journey) => {
-  const duration = getJourneyDuration(journey);
-  if (duration === STRINGS.UNKNOWN_TEXT) {
-    return STRINGS.UNKNOWN_TEXT;
-  }
-  const time = getFlightTimeObject(duration);
-  if (!time.h && !time.m) {
-    return STRINGS.UNKNOWN_TEXT;
-  }
-  return `${time.h}h ${time.m}m`;
-};
-
 const getByIataCode = (iataCode) => {
   return airports.findWhere({ iata: iataCode });
 };
@@ -336,10 +281,6 @@ const getCityByIataCode = (iataCode) => {
     return STRINGS.UNKNOWN_TEXT;
   }
   return city;
-};
-
-const getRoute = (journey) => {
-  return journey?.route;
 };
 
 const toRoute = (route) => {
@@ -370,39 +311,6 @@ const getFlight = (targetTask) => {
   return null;
 };
 
-const hasJourney = (targetTask) => {
-  return !!targetTask?.movement?.journey;
-};
-
-const getJourney = (targetTask) => {
-  if (hasJourney(targetTask)) {
-    return targetTask.movement.journey;
-  }
-  return null;
-};
-
-const getDepartureTime = (journey) => {
-  return journey?.departure?.time;
-};
-
-const getArrivalTime = (journey) => {
-  return journey?.arrival?.time;
-};
-
-const toFormattedDepartureDateTime = (journey, dateFormat = DATE_FORMATS.LONG) => {
-  if (!journey?.departure?.time) {
-    return STRINGS.UNKNOWN_TEXT;
-  }
-  return getFormattedDate(journey.departure.time, dateFormat);
-};
-
-const toFormattedArrivalDateTime = (journey, dateFormat = DATE_FORMATS.LONG) => {
-  if (!journey?.arrival?.time) {
-    return STRINGS.UNKNOWN_TEXT;
-  }
-  return getFormattedDate(journey.arrival.time, dateFormat);
-};
-
 const toFormattedLocation = (location) => {
   if (!location) {
     return STRINGS.UNKNOWN_TEXT;
@@ -411,27 +319,13 @@ const toFormattedLocation = (location) => {
   return airport ? `${airport.get('city')}, ${airport.get('country')}` : STRINGS.UNKNOWN_TEXT;
 };
 
-const getDepartureLocation = (journey) => {
-  if (!journey?.departure?.location) {
-    return STRINGS.UNKNOWN_TEXT;
-  }
-  return journey.departure.location;
-};
-
-const getArrivalLocation = (journey) => {
-  if (!journey?.arrival?.location) {
-    return STRINGS.UNKNOWN_TEXT;
-  }
-  return journey.arrival.location;
-};
-
 /**
  * If country code is not provided, it will use the arrival location
  * to extract the arrival country code.
  */
 const getItineraryArrivalCountryCode = (itinerary) => {
   if (!itinerary?.arrival?.country) {
-    const arrivalLoc = getArrivalLocation(itinerary);
+    const arrivalLoc = JourneyUtil.arrivalLoc(itinerary);
     if (!arrivalLoc) {
       return STRINGS.UNKNOWN_TEXT;
     }
@@ -453,7 +347,7 @@ const getItineraryArrivalCountryCode = (itinerary) => {
  */
 const getItineraryDepartureCountryCode = (itinerary) => {
   if (!itinerary?.departure?.country) {
-    const departureLoc = getDepartureLocation(itinerary);
+    const departureLoc = JourneyUtil.departureLoc(itinerary);
     if (!departureLoc) {
       return STRINGS.UNKNOWN_TEXT;
     }
@@ -542,8 +436,8 @@ const toAirlineName = (airlineCode, refDataAirlineCodes) => {
 };
 
 const toItineraryRelativeTime = (index, itinerary, itineraries) => {
-  const previousLegArrivalTime = getArrivalTime(itineraries[index - 1]);
-  const nextLegDepartureTime = getDepartureTime(itinerary);
+  const previousLegArrivalTime = JourneyUtil.arrivalTime((itineraries[index - 1]));
+  const nextLegDepartureTime = JourneyUtil.departureTime((itinerary));
   return (
     <div className="font__light">
       {calculateTimeDifference(
@@ -563,18 +457,9 @@ const MovementUtil = {
   outcomeStatusTag: getOutcomeStatusTag,
   outcomeFLO: getOutcomeFrontlineOfficer,
   outcomeFLOArrests: getOutcomeFrontlineArrests,
-  movementRoute: getRoute,
   convertMovementRoute: toRoute,
   seatNumber: getSeatNumber,
   movementFlight: getFlight,
-  movementJourney: getJourney,
-  departureTime: getDepartureTime,
-  direction: getMovementDirection,
-  arrivalTime: getArrivalTime,
-  formatDepartureTime: toFormattedDepartureDateTime,
-  formatArrivalTime: toFormattedArrivalDateTime,
-  departureLoc: getDepartureLocation,
-  arrivalLoc: getArrivalLocation,
   flightNumber: getFlightNumber,
   airlineOperator: getAirlineOperator,
   status: getDepartureStatus,
@@ -582,13 +467,9 @@ const MovementUtil = {
   description: toDescriptionText,
   airlineName: toAirlineName,
   formatLoc: toFormattedLocation,
-  flightDuration: getJourneyDuration,
-  flightTimeObject: getFlightTimeObject,
-  formatFlightTime: toFormattedFlightTime,
   movementMode: getMovementMode,
   iconDescription: getIconDescription,
   modeIcon: getMovementModeIcon,
-  movementItinerary: getMovementItinerary,
   itinFlightNumber: getItineraryFlightNumber,
   itinDepartureCountryCode: getItineraryDepartureCountryCode,
   itinArrivalCountryCode: getItineraryArrivalCountryCode,
@@ -617,14 +498,10 @@ const MovementUtil = {
 export default MovementUtil;
 
 export {
-  getRoute,
   toRoute,
   getBusinessKey,
   getSeatNumber,
   getFlight,
-  getJourney,
-  getDepartureTime,
-  getArrivalTime,
   getNamedPassenger,
   hasCarrierCounts,
   hasCheckinDate,
@@ -639,10 +516,6 @@ export {
   isSinglePassenger,
   isTaskDetailsPassenger,
   isValid,
-  toFormattedDepartureDateTime,
-  toFormattedArrivalDateTime,
-  getDepartureLocation,
-  getArrivalLocation,
   getFlightNumber,
   getAirlineOperator,
   getDepartureStatus,
@@ -650,9 +523,6 @@ export {
   toDescriptionText,
   toAirlineName,
   toFormattedLocation,
-  getJourneyDuration,
-  toFormattedFlightTime,
-  getFlightTimeObject,
   getItineraryFlightNumber,
   getItineraryDepartureCountryCode,
   getItineraryArrivalCountryCode,
@@ -661,6 +531,5 @@ export {
   getUpdatedStatus,
   getCityByIataCode,
   toVoyageText,
-  getMovementDirection,
   getMovementModeIcon,
 };
